+++ src/game/player/elements/menu/type.ts
@@ -1,6 +1,6 @@
-import type {Sentence} from "@core/elements/text";
 import type {Choice} from "@core/elements/menu";
 import type {GameState} from "@player/gameState";
+import {Sentence} from "@core/elements/character/sentence";
 
 export interface MenuElementProps {
     prompt: Sentence;